requirements.txt
discord.py[voice]==2.3.2
flask==3.1.0
PyNaCl==1.5.0
